added new validation, response http and another thing

+++ Controllers/AdmController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using blog.utils.Filters.FiltersDTO;
 using blog.utils.Filters.FiltersQuerys;
+using blog.utils.Responses;
 using Blog.entities;
 using Blog.SetUnitOfWork;
 using Blog.utils;
@@ -37,106 +38,240 @@ namespace blog.Controllers
         [Authorize(Roles = "AdminRole, SuperAdminRole")]
         public async Task<IActionResult> DeletePost(long postId)
         {
-            PostEntity post = await this._uow.PostRepository.Get(postId);
-            ApplicationUser user = await this._uow.UserRepository.Get(post.ApplicationUserId);
-            await this._uow.PostRepository.Delete(post, user);
+            if (postId <= 0)
+            {
+                return BadRequest(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 400,
+                    Message = "Post id is required",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
 
-            return Ok(new Response(
-                "success",
-                "Post deleted with successfully",
-                201,
-                null
-            ));
+            PostEntity? post = await _uow.PostRepository.Get(postId);
+
+            if (post == null)
+            {
+                return NotFound(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 404,
+                    Message = "Post not found",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
+
+            ApplicationUser? user = await _uow.UserRepository.Get(post.ApplicationUserId);
+
+            if (user == null)
+            {
+                return NotFound(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 400,
+                    Message = "User not found",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
+
+            await _uow.PostRepository.Delete(post, user);
+
+            return Ok(new ResponseBody<string>
+            {
+                Body = null,
+                Code = 200,
+                Datetime = DateTimeOffset.Now,
+                Message = "Post deleted",
+                Status = true
+            });
         }
 
         [HttpDelete("delete-comment/{commentId:required}")]
         [Authorize(Roles = "AdminRole, SuperAdminRole")]
         public async Task<IActionResult> DeleteComment(ulong commentId)
         {
-            CommentEntity comment = await this._uow.CommentRepository.Get(commentId);
+            CommentEntity? comment = await this._uow.CommentRepository.Get(commentId);
+
+            if (comment == null)
+            {
+                return NotFound(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 400,
+                    Message = "Comment not found",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
+
             await this._uow.CommentRepository.Delete(comment);
 
-            return Ok(new Response(
-                "success",
-                "Comment deleted with successfully",
-                201,
-                null
-            ));
+            return Ok(new ResponseBody<string>
+            {
+                Body = null,
+                Code = 200,
+                Datetime = DateTimeOffset.Now,
+                Message = "comment deleted",
+                Status = true
+            });
         }
 
         [HttpPost("{userId:required}")]
         [Authorize(Roles = "SuperAdminRole")]
         public async Task<IActionResult> AddRoleAdmInUser(string userId)
         {
-            ApplicationUser user = await _uow.UserRepository.Get(userId);
+            if (string.IsNullOrWhiteSpace(userId))
+            {
+                return BadRequest(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 400,
+                    Message = "Id is required",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
+
+            ApplicationUser? user = await _uow.UserRepository.Get(userId);
+
+            if (user == null)
+            {
+                return NotFound(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 404,
+                    Message = "User not found",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
+
             IdentityRole? role = await _roleManager.FindByNameAsync("AdminRole");
 
             if (role == null || string.IsNullOrEmpty(role.Name))
-                return Ok(new ResponseException(
-                    "Role not found",
-                    StatusCodes.Status404NotFound
-                ));
+                return NotFound(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 404,
+                    Message = "Role not found",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
 
             if (await _userManager.IsInRoleAsync(user, role.Name))
-                return BadRequest(new ResponseException($"User '{user.UserName}' does have the '{role.Name}' role.", 400));
+                return BadRequest(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 400,
+                    Message = $"User '{user.UserName}' have the '{role.Name}' role.",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
 
             var result = await _userManager.AddToRoleAsync(user, role.Name);
 
             if (!result.Succeeded)
             {
-                var errors = result.Errors.Select(e => e.Description);
-                return BadRequest(new ResponseException(
-                    "Error removing AdminRole from user.",
-                    StatusCodes.Status500InternalServerError,
-                    "fail",
-                    errors
-                ));
+                IEnumerable<string>? errors = result.Errors.Select(e => e.Description);
+                return StatusCode(StatusCodes.Status500InternalServerError, new ResponseBody<IEnumerable<string>>
+                {
+                    Message = "Error adding AdminRole from user.",
+                    Code = StatusCodes.Status500InternalServerError,
+                    Status = false,
+                    Body = errors,
+                    Datetime = DateTimeOffset.Now
+                });
             }
 
-            return Ok(new Response(
-                "success",
-                $"The user {user.UserName} now is a adm!!!!! ",
-                StatusCodes.Status201Created,
-                null
-            ));
+            return Ok(new ResponseBody<string>
+            {
+                Message = $"The user {user.UserName} now is a adm!!!!! ",
+                Code = StatusCodes.Status200OK,
+                Body = null,
+                Datetime = DateTimeOffset.Now,
+                Status = false
+            });
         }
 
         [HttpDelete("remove-role-adm/{userId:required}")]
         [Authorize(Roles = "SuperAdminRole")]
         public async Task<IActionResult> RemoveRoleAdmInUser(string userId)
         {
-            ApplicationUser user = await _uow.UserRepository.Get(userId);
+            if (string.IsNullOrWhiteSpace(userId))
+            {
+                return BadRequest(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 400,
+                    Message = "Id is required",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
+
+            ApplicationUser? user = await _uow.UserRepository.Get(userId);
+
+            if (user == null)
+            {
+                return NotFound(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 404,
+                    Message = "User not found",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
+
             IdentityRole? role = await _roleManager.FindByNameAsync("AdminRole");
 
             if (role == null || string.IsNullOrEmpty(role.Name))
-                return Ok(new ResponseException(
-                    "Role not found",
-                    StatusCodes.Status404NotFound
-                ));
+                return NotFound(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 404,
+                    Message = "Role not found",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
 
             if (!await _userManager.IsInRoleAsync(user, role.Name))
-                return BadRequest(new ResponseException($"User '{user.UserName}' does not have the '{role.Name}' role.", 400));
-            
+                return BadRequest(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 400,
+                    Message = $"User '{user.UserName}' does have the '{role.Name}' role.",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
 
             var result = await _userManager.RemoveFromRoleAsync(user, role.Name);
 
             if (!result.Succeeded)
             {
                 var errors = result.Errors.Select(e => e.Description);
-                return BadRequest(new ResponseException(
-                    "Error removing AdminRole from user.",
-                    StatusCodes.Status500InternalServerError,
-                    "fail",
-                    errors
-                ));
+                return StatusCode(StatusCodes.Status500InternalServerError, new ResponseBody<IEnumerable<string>>
+                {
+                    Message = "Error removing AdminRole from user.",
+                    Code = StatusCodes.Status500InternalServerError,
+                    Status = false,
+                    Body = errors,
+                    Datetime = DateTimeOffset.Now
+                });
             }
 
-            return Ok(new Response(
-                "success",
-                $"Role adm removed of user {user.UserName}!",
-                StatusCodes.Status201Created,
-                null
-            ));
+            return Ok(new ResponseBody<string>
+            {
+                Body = null,
+                Message = $"Role: {role.Name} adm removed of user {user.UserName}!",
+                Code = StatusCodes.Status200OK,
+                Datetime = DateTimeOffset.Now,
+                Status = false
+            });
         }
 
         [HttpGet]
@@ -146,44 +281,74 @@ namespace blog.Controllers
             IdentityRole? role = await _roleManager.FindByNameAsync("AdminRole");
 
             if (role == null || string.IsNullOrEmpty(role.Name))
-                return Ok(new ResponseException(
-                    "Role not found",
-                    StatusCodes.Status404NotFound
-                ));
+                return NotFound(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 404,
+                    Message = "Role not found",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
 
             IList<ApplicationUser>? result = await _userManager.GetUsersInRoleAsync(role.Name);
 
-            return Ok(new Response(
-                "success",
-                "All user has role adm listed!",
-                StatusCodes.Status201Created,
-                result
-            ));
+            return Ok(new ResponseBody<IList<ApplicationUser>>
+            {
+                Status = true,
+                Message = "All user has role adm listed!",
+                Code = StatusCodes.Status200OK,
+                Body = result,
+                Datetime = DateTimeOffset.Now
+            });
         }
 
         [HttpDelete("delete-user/{userId:required}")]
         [Authorize(Roles = "SuperAdminRole")]
-        public async Task<IActionResult> DeleteUser(string userId) 
+        public async Task<IActionResult> DeleteUser(string userId)
         {
-            ApplicationUser user = await _uow.UserRepository.Get(userId);
+            if (string.IsNullOrWhiteSpace(userId))
+            {
+                return BadRequest(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 400,
+                    Message = "Id is required",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
+
+            ApplicationUser? user = await _uow.UserRepository.Get(userId);
+
+            if (user == null)
+            {
+                return NotFound(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 404,
+                    Message = "User not found",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
+
             await _uow.UserRepository.Delete(user);
 
-            return Ok(new Response(
-                "success",
-                $"The user {user.UserName} deleted! ",
-                StatusCodes.Status201Created,
-                null
-            ));
+            return Ok(new ResponseBody<string>
+            {
+                Status = true,
+                Message = $"The user {user.UserName} was deleted!",
+                Code = StatusCodes.Status200OK,
+                Body = null,
+                Datetime = DateTimeOffset.Now
+            });
         }
 
-        [HttpGet("get-all-category/{canFilter:bool?}")]
+        [HttpGet("get-all-category")]
         [Authorize(Roles = "AdminRole, SuperAdminRole")]
-        public async Task<IActionResult> GetAllCategory([FromQuery] CategoryFilterDTO filter, bool canFilter = false)
+        public async Task<IActionResult> GetAllCategory([FromQuery] CategoryFilterDTO filter)
         {
             IQueryable<CategoryEntity> query = _uow.CategoryRepository.GetAll();
-
-            if (canFilter == true)
-                query = CategoryQueryFilter.ApplyFilters(query, filter);
 
             PaginatedList<CategoryEntity> result = await PaginatedList<CategoryEntity>.CreateAsync(query, filter.PageNumber, filter.PageSize);
