+++ Controllers/UserConfigController.cs
@@ -5,6 +5,7 @@ using System.Security.Claims;
 using System.Threading.Tasks;
 using blog.DTOs.UserConfig;
 using blog.entities;
+using blog.utils.Responses;
 using Blog.entities;
 using Blog.SetUnitOfWork;
 using Blog.utils;
@@ -33,15 +34,52 @@ namespace blog.Controllers
         public async Task<IActionResult> Get()
         {
             string? userId = User.FindFirst(ClaimTypes.Sid)?.Value;
-            ApplicationUser user = await _uow.UserRepository.Get(userId);
-            UserConfigEntity config = await _uow.UserConfigRepository.GetAsync(user);
+            if (string.IsNullOrWhiteSpace(userId))
+            {
+                return Unauthorized(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 401,
+                    Message = "You are not authorizetion",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
 
-            return Ok(new Response(
-                "success",
-                "User config founded!!!",
-                200,
-                config
-            ));
+            ApplicationUser? user = await _uow.UserRepository.Get(userId);
+            if (user == null)
+            {
+                return NotFound(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 404,
+                    Message = "User not found",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
+
+            UserConfigEntity? config = await _uow.UserConfigRepository.GetAsync(user);
+            if (config == null)
+            {
+                return NotFound(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 404,
+                    Message = "Config User not found",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
+
+            return Ok(new ResponseBody<UserConfigEntity>
+            {
+                Status = true,
+                Message = "User config founded!!!",
+                Code = 200,
+                Body = config,
+                Datetime = DateTimeOffset.Now
+            });
         }
 
         [HttpDelete]
@@ -49,16 +87,55 @@ namespace blog.Controllers
         public async Task<IActionResult> Delete()
         {
             string? userId = User.FindFirst(ClaimTypes.Sid)?.Value;
-            ApplicationUser user = await _uow.UserRepository.Get(userId);
-            UserConfigEntity config = await _uow.UserConfigRepository.GetAsync(user);
+            if (string.IsNullOrWhiteSpace(userId))
+            {
+                return Unauthorized(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 401,
+                    Message = "You are not authorizetion",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
+
+            ApplicationUser? user = await _uow.UserRepository.Get(userId);
+
+            if (user == null)
+            {
+                return NotFound(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 404,
+                    Message = "User not found",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
+
+            UserConfigEntity? config = await _uow.UserConfigRepository.GetAsync(user);
+            if (config == null)
+            {
+                return NotFound(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 404,
+                    Message = "Config User not found",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
+            
             await _uow.UserConfigRepository.DeleteAsync(config);
 
-            return Ok(new Response(
-                "success",
-                "User config deleted!!!",
-                200,
-                null
-            ));
+            return Ok(new ResponseBody<UserConfigEntity>
+            {
+                Status = true,
+                Message = "User config deleted!!!",
+                Code = 200,
+                Body = null,
+                Datetime = DateTimeOffset.Now
+            });
         }
 
         [HttpPost]
@@ -69,15 +146,54 @@ namespace blog.Controllers
                 return BadRequest(ModelState);
 
             string? userId = User.FindFirst(ClaimTypes.Sid)?.Value;
-            ApplicationUser user = await _uow.UserRepository.Get(userId);
+            if (string.IsNullOrWhiteSpace(userId))
+            {
+                return Unauthorized(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 401,
+                    Message = "You are not authorizetion",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
+
+            bool check = await _uow.UserConfigRepository.Exists(userId);
+            if (check)
+            {
+                return Conflict(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 409,
+                    Message = "User already have a config",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
+
+            ApplicationUser? user = await _uow.UserRepository.Get(userId);
+            if (user == null)
+            {
+                return NotFound(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 404,
+                    Message = "User not found",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
+
             UserConfigEntity config = await _uow.UserConfigRepository.CreateAsync(dto, user);
 
-            return Ok(new Response(
-                "success",
-                "User config created!!!",
-                200,
-                config
-            ));
+            return Ok(new ResponseBody<UserConfigEntity>
+            {
+                Status = true,
+                Message = "User config created!!!",
+                Code = 200,
+                Body = config,
+                Datetime = DateTimeOffset.Now
+            });
         }
 
         [HttpPut]
@@ -86,17 +202,56 @@ namespace blog.Controllers
         {
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
-                
-            string? userId = User.FindFirst(ClaimTypes.Sid)?.Value;
-            ApplicationUser user = await _uow.UserRepository.Get(userId);
-            UserConfigEntity config = await _uow.UserConfigRepository.UpdateAsync(dto, user);
 
-            return Ok(new Response(
-                "success",
-                "User config updated!!!",
-                200,
-                config
-            ));
+            string? userId = User.FindFirst(ClaimTypes.Sid)?.Value;
+            if (string.IsNullOrWhiteSpace(userId))
+            {
+                return Unauthorized(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 401,
+                    Message = "You are not authorizetion",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
+
+            ApplicationUser? user = await _uow.UserRepository.Get(userId);
+            if (user == null)
+            {
+                return NotFound(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 404,
+                    Message = "User not found",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
+
+            UserConfigEntity? config = await _uow.UserConfigRepository.GetAsync(user);
+            if (config == null)
+            {
+                return NotFound(new ResponseBody<string>
+                {
+                    Body = null,
+                    Code = 404,
+                    Message = "Config User not found",
+                    Status = false,
+                    Datetime = DateTimeOffset.Now
+                });
+            }
+
+            UserConfigEntity configUpdated = await _uow.UserConfigRepository.UpdateAsync(dto, user, config);
+
+            return Ok(new ResponseBody<UserConfigEntity>
+            {
+                Status = true,
+                Message = "User config updated!!!",
+                Code = 200,
+                Body = configUpdated,
+                Datetime = DateTimeOffset.Now
+            });
         }
 
     }
